+++ src/Layout.js
@@ -1,15 +1,10 @@
 import React, { Component } from 'react';
-import classNames from 'classnames';
 import { withStyles } from 'material-ui/styles';
-import Typography from 'material-ui/Typography';
-import Paper from '@material-ui/core/Paper';
-import Input from '@material-ui/core/Input';
-import Avatar from 'material-ui/Avatar';
 import Sidebar from './components/Sidebar.jsx';
 import ChatHeader from './components/ChatHeader.jsx';
+import Chat from './components/Chat.jsx';
 
 import { chats, messages } from './mock-data.json';
-import titleInitials from './utils/title-initials'
 
 const styles = theme => ({
   root: {
@@ -18,56 +13,6 @@ const styles = theme => ({
     width: '100%',
     height: '100%',
     backgroundColor: theme.palette.background.default,
-  },
-  chatLayout: {
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    height: '100%',
-    overflow: 'hidden',
-  },
-  messagesWrapper: {
-    overflowX: 'scroll',
-    height: '100%',
-    width: '100%',
-    paddingTop: theme.spacing.unit * 3,
-    paddingBottom: '120px',
-  },
-  messageInputWrapper: {
-    position: 'fixed',
-    left: 'auto',
-    right: 0,
-    bottom: 0,
-    width: `calc(100% - 320px)`,
-    padding: theme.spacing.unit * 3,
-  },
-  messageInput: {
-    padding: theme.spacing.unit * 2,
-  },
-  messageWrapper: {
-    display: 'flex',
-    justifyContent: 'flex-start',
-    alignItems: 'center',
-    padding: `${theme.spacing.unit}px ${theme.spacing.unit * 3}px`,
-  },
-  messageWrappperFromMe: {
-    justifyContent: 'flex-end',
-  },
-  message: {
-    maxWidth: '70%',
-    minWidth: '10%',
-    padding: theme.spacing.unit,
-    marginLeft: theme.spacing.unit * 2,
-  },
-  messageFromMe: {
-    marginRight: theme.spacing.unit * 2,
-    backgroundColor: '#e6dcff'
-  },
-  toolbar: theme.mixins.toolbar,
-  content: {
-    flexGrow: 1,
-    backgroundColor: theme.palette.background.default,
-    padding: theme.spacing.unit * 3
   },
 });
 
@@ -100,40 +45,7 @@ class Layout extends Component {
       <div className={classes.root}>
         <ChatHeader />
         <Sidebar chats={chats}/>
-        <main className={classes.chatLayout} ref={this.refMessagesWrapper}>
-          <div className={classes.messagesWrapper}>
-          {
-            messages && messages.map((message, index) => {
-              const isMessageFromMe = message.sender === 'me';
-
-              const userAvatar = (
-                <Avatar> {titleInitials(message.sender[0])} </Avatar>
-              );
-
-              return (
-                 <div key={index}
-                 className={classNames(classes.messageWrapper, isMessageFromMe && classes.messageWrappperFromMe)}>
-                {!isMessageFromMe && userAvatar}
-                <Paper className={classNames(classes.message, isMessageFromMe && classes.messageFromMe)}>
-                  <Typography variant="caption">
-                    {message.sender}
-                  </Typography>
-                  <Typography variant="body1">
-                    {message.content}
-                  </Typography>
-                </Paper>
-                {isMessageFromMe && userAvatar}
-              </div>
-              );
-            })
-          }
-          </div>
-          <div className={classes.messageInputWrapper}>
-            <Paper className={classes.messageInput} elevation={6}>
-              <Input fullWidth placeholder="Type your message…"/>
-            </Paper>
-          </div>
-        </main>
+        <Chat messages={messages}/>
       </div>
     );
   }
