+++ src/components/Chat.jsx
@@ -1,13 +1,105 @@
-import React, { PureComponent } from 'react';
+import React, { Component } from 'react';
+import classNames from 'classnames';
+import { withStyles } from 'material-ui/styles';
+import Typography from 'material-ui/Typography';
+import Paper from '@material-ui/core/Paper';
+import Input from '@material-ui/core/Input';
+import Avatar from 'material-ui/Avatar';
+import titleInitials from '../utils/title-initials';
 
-class Chat extends PureComponent {
+const styles = theme => ({
+  chatLayout: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    height: '100%',
+    overflow: 'hidden',
+  },
+  messagesWrapper: {
+    overflowX: 'scroll',
+    height: '100%',
+    width: '100%',
+    paddingTop: theme.spacing.unit * 3,
+    paddingBottom: '120px',
+  },
+  messageInputWrapper: {
+    position: 'fixed',
+    left: 'auto',
+    right: 0,
+    bottom: 0,
+    width: `calc(100% - 320px)`,
+    padding: theme.spacing.unit * 3,
+  },
+  messageInput: {
+    padding: theme.spacing.unit * 2,
+  },
+  messageWrapper: {
+    display: 'flex',
+    justifyContent: 'flex-start',
+    alignItems: 'center',
+    padding: `${theme.spacing.unit}px ${theme.spacing.unit * 3}px`,
+  },
+  messageWrappperFromMe: {
+    justifyContent: 'flex-end',
+  },
+  message: {
+    maxWidth: '70%',
+    minWidth: '10%',
+    padding: theme.spacing.unit,
+    marginLeft: theme.spacing.unit * 2,
+  },
+  messageFromMe: {
+    marginRight: theme.spacing.unit * 2,
+    backgroundColor: '#e6dcff'
+  },
+  toolbar: theme.mixins.toolbar,
+  content: {
+    flexGrow: 1,
+    backgroundColor: theme.palette.background.default,
+    padding: theme.spacing.unit * 3
+  },
+});
 
+class Chat extends Component {
   render() {
+    const { classes, messages } = this.props;
     return (
-      <div></div>
+      <main className={classes.chatLayout} ref={this.refMessagesWrapper}>
+        <div className={classes.messagesWrapper}>
+        {
+          messages && messages.map((message, index) => {
+            const isMessageFromMe = message.sender === 'me';
+
+            const userAvatar = (
+              <Avatar> {titleInitials(message.sender[0])} </Avatar>
+            );
+
+            return (
+               <div key={index}
+               className={classNames(classes.messageWrapper, isMessageFromMe && classes.messageWrappperFromMe)}>
+              {!isMessageFromMe && userAvatar}
+              <Paper className={classNames(classes.message, isMessageFromMe && classes.messageFromMe)}>
+                <Typography variant="caption">
+                  {message.sender}
+                </Typography>
+                <Typography variant="body1">
+                  {message.content}
+                </Typography>
+              </Paper>
+              {isMessageFromMe && userAvatar}
+            </div>
+            );
+          })
+        }
+        </div>
+        <div className={classes.messageInputWrapper}>
+          <Paper className={classes.messageInput} elevation={6}>
+            <Input fullWidth placeholder="Type your message…"/>
+          </Paper>
+        </div>
+      </main>
     );
   }
-
 }
 
-export default Chat;
+export default withStyles(styles)(Chat);
